% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallOptionsOverTime.R
\name{CallOptionsOverTime}
\alias{CallOptionsOverTime}
\title{Time Sensitivity Analysis for European Call Option}
\usage{
CallOptionsOverTime(
  stock_data,
  future_times,
  strike_price,
  max.p = 5,
  max.q = 5
)
}
\arguments{
\item{stock_data}{Numeric vector representing the historical stock prices.}

\item{future_times}{Numeric vector representing the future time periods (in the same units as stock_data) for which the option value is calculated.}

\item{strike_price}{Numeric value representing the fixed strike price to analyze.}

\item{max.p}{The maximum order of the autoregressive part of the ARMA model (default is 5).}

\item{max.q}{The maximum order of the moving average part of the ARMA model (default is 5).}
}
\value{
A data frame containing the future times and their corresponding option values.
}
\description{
This function analyzes the sensitivity of the European call option value to changes in the time to maturity.
It calculates the option value for a fixed strike price over a range of future time periods using the `europeanCallOptionValue` function.
}
\examples{
\dontrun{
stock_data <- c(100, 102, 105, 107, 110, 108, 105, 103, 101, 99, 98, 97)
future_times <- c(1, 3, 5)
strike_price <- 110
CallOptionsOverTime(stock_data, future_times, strike_price)
}

}
