% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_expectation_normal.R
\name{conditional_expectation_normal}
\alias{conditional_expectation_normal}
\title{Calculate Conditional Expectation for Normal Distribution}
\usage{
conditional_expectation_normal(mu, sigma, S)
}
\arguments{
\item{mu}{Mean of the normal distribution.}

\item{sigma}{Standard deviation of the normal distribution.}

\item{S}{Threshold value.}
}
\value{
The conditional expectation \eqn{E[X | X > S]}.
}
\description{
This function calculates the conditional expectation \eqn{E[X | X > S]} for a normal distribution
with mean \eqn{\mu} and standard deviation \eqn{\sigma}, given a threshold \eqn{S}.
}
\examples{
conditional_expectation_normal(0, 1, 1.5)
conditional_expectation_normal(10, 2, 12)

}
