% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_expectation_normal_negation.R
\name{conditional_expectation_normal_negation}
\alias{conditional_expectation_normal_negation}
\title{Calculate Conditional Expectation for Normal Distribution Below a Threshold}
\usage{
conditional_expectation_normal_negation(mu, sigma, S)
}
\arguments{
\item{mu}{Mean of the normal distribution.}

\item{sigma}{Standard deviation of the normal distribution.}

\item{S}{Threshold value.}
}
\value{
The conditional expectation \eqn{E[X | X < S]}.
}
\description{
This function calculates the conditional expectation \eqn{E[X | X < S]} for a normal distribution
with mean \eqn{\mu} and standard deviation \eqn{\sigma}, given a threshold \eqn{S}.
}
\examples{
conditional_expectation_normal_negation(0, 1, 1.5)
conditional_expectation_normal_negation(10, 2, 12)

}
