% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/europeanCallOptionValue.R
\name{europeanCallOptionValue}
\alias{europeanCallOptionValue}
\title{Calculate European Call Option Value}
\usage{
europeanCallOptionValue(
  stock_data,
  future_time,
  sell_value,
  max.p = 5,
  max.q = 5
)
}
\arguments{
\item{stock_data}{Numeric vector representing the historical stock prices.}

\item{future_time}{Future time period (in same units as stock_data) for which the put option value is calculated.}

\item{sell_value}{The strike price of the European call option.}

\item{max.p}{The maximum order of the autoregressive part of the ARMA model.}

\item{max.q}{The maximum order of the moving average part of the ARMA model.}
}
\value{
The expected payoff of the European call option.
}
\description{
This function calculates the value of a European call option based on historical stock prices, a future time period, and a specified strike price.
}
\examples{
\dontrun{
stock_data <- c(100, 102, 105, 107, 110)
future_time <- 5
sell_value <- 115
europeanCallOptionValue(stock_data, future_time, sell_value)
}

}
