% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/europeanPutOptionValue.R
\name{europeanPutOptionValue}
\alias{europeanPutOptionValue}
\title{Calculate European Put Option Value}
\usage{
europeanPutOptionValue(stock_data, future_time, sell_value, p = 5, q = 5)
}
\arguments{
\item{stock_data}{Numeric vector representing the historical stock prices.}

\item{future_time}{Future time period (in same units as stock_data) for which the put option value is calculated.}

\item{sell_value}{The strike price of the European put option.}

\item{p}{The maximum order of the autoregressive part of the ARMA model.}

\item{q}{The maximum order of the moving average part of the ARMA model.}
}
\value{
An estimate of the value of a European put option.
}
\description{
This function calculates the value of a European put option based on historical stock prices, a future time period, and a specified strike price.
}
\examples{
\dontrun{
stock_data <- c(100, 102, 105, 107, 110)
future_time <- 5
sell_value <- 115
europeanPutOptionValue(stock_data, future_time, sell_value)
}

}
