% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PutOptionsOverStrikePrices.R
\name{PutOptionsOverStrikePrices}
\alias{PutOptionsOverStrikePrices}
\title{Strike Price Sensitivity Analysis for European Put Option}
\usage{
PutOptionsOverStrikePrices(
  stock_data,
  future_time,
  sell_values,
  max.p = 5,
  max.q = 5,
  method = "CSS-ML"
)
}
\arguments{
\item{stock_data}{Numeric vector of stock prices data.}

\item{future_time}{Numeric constant of the future time}

\item{max.p}{The maximum order of the autoregressive part of the ARMA model (default is 5).}

\item{max.q}{The maximum order of the moving average part of the ARMA model (default is 5).}

\item{method}{The way that the ARMA model is calculated}

\item{strike_prices}{Numeric vector representing the range of strike prices to analyze.}
}
\value{
Estimated values of a European put option at different sell_value
}
\description{
This function analyzes the sensitivity of the European put option value to changes in the strike price for a single future time period.
It calculates the option value for a range of strike prices using the `europeanPutOptionValue` function.
}
\examples{
\dontrun{
n <- 100
set.seed(42)
arma_values <- arima.sim(n = n, model = list(ar = c(0.6), ma = c(0.5, -0.5)))
linear_model <- 5 + 1:n
stock_data <- arma_values + linear_model

future_time <- 2
sell_values <- seq(90, 110, length.out = 5)

# Run the function
PutOptionsOverStrikePrices(stock_data, future_time, sell_values)
}

}
