% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallOptionsOverTime.R
\name{CallOptionsOverTime}
\alias{CallOptionsOverTime}
\title{Time Sensitivity Analysis for European Call Option}
\usage{
CallOptionsOverTime(stock_data, future_times, buy_value, max.p = 5, max.q = 5)
}
\arguments{
\item{stock_data}{Numeric vector of stock prices data.}

\item{future_times}{Numeric vector of the future times}

\item{max.p}{The maximum order of the autoregressive part of the ARMA model (default is 5).}

\item{max.q}{The maximum order of the moving average part of the ARMA model (default is 5).}

\item{strike_price}{Numeric value representing the fixed strike price to analyze.}
}
\value{
Estimated values of a European call option at different future_times
}
\description{
This function analyzes the sensitivity of the European call option value to changes in the time to maturity.
It calculates the option value for a fixed strike price over a range of future time periods using the `europeanCallOptionValue` function.
}
\examples{
\dontrun{
library(stats)
library(forecast)

# Create simulated data
n <- 100
set.seed(42)
arma_values <- arima.sim(n = n, model = list(ar = c(0.6), ma = c(0.5, -0.5)))
linear_model <- 5 + 1:n
stock_data <- arma_values + linear_model

future_times <- seq(1, 10, length.out = 10)
buy_value <- 100

# Run the function
CallOptionsOverTime(stock_data, future_times, buy_value)
}

}
