% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PutOptionsOverStrikePrices.R
\name{PutOptionsOverStrikePrices}
\alias{PutOptionsOverStrikePrices}
\title{Strike Price Sensitivity Analysis for European Put Option}
\usage{
PutOptionsOverStrikePrices(
  stock_data,
  future_time,
  sell_values,
  max.p = 5,
  max.q = 5,
  method = "CSS-ML"
)
}
\arguments{
\item{stock_data}{Numeric vector of stock prices data.}

\item{future_time}{Numeric constant of the future time}

\item{sell_values}{Numeric vector of the sell values to calculate the put option values at}

\item{max.p}{The maximum order of the autoregressive part of the ARMA model (default is 5)}

\item{max.q}{The maximum order of the moving average part of the ARMA model (default is 5)}

\item{method}{The way that the ARMA model is calculated, accepted values are "ML", "CSS-ML" and "CSS"}
}
\value{
Estimated values of a European put option at different sell_value
}
\description{
This function analyzes the sensitivity of the European put option value to changes in the strike price for a single future time period.
It calculates the option value for a range of strike prices using the `europeanPutOptionValue` function.
}
