% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackPutOptionValueDifferences.R
\name{trackPutOptionValueDifferences}
\alias{trackPutOptionValueDifferences}
\title{Track European Put Option Value Differences}
\usage{
trackPutOptionValueDifferences(closing_prices, steps, iter, p = 5, q = 5)
}
\arguments{
\item{closing_prices}{Numeric vector representing the historical closing prices of the stock.}

\item{steps}{The number of steps (time periods) to look ahead for the option value calculation.}

\item{iter}{The total number of iterations to perform.}

\item{p}{The maximum order of the autoregressive part of the ARMA model (default is 5).}

\item{q}{The maximum order of the moving average part of the ARMA model (default is 5).}
}
\value{
A list containing:
\describe{
  \item{tracker}{Numeric vector of estimated European put option values.}
  \item{actual_values}{Numeric vector of actual European put option values.}
  \item{difference}{Numeric vector of differences between actual and estimated values.}
}
}
\description{
This function tracks the differences between the estimated European put option values and the actual values over a specified number of iterations.
}
\examples{
\dontrun{
closing_prices <- c(100, 102, 105, 107, 110, 108, 105, 103, 101, 99, 98, 97)
steps <- 2
iter <- 10
result <- trackPutOptionValueDifferences(closing_prices, steps, iter)
}

}
